refactor: move skills to pkg/skills; name the capability concept in docs (#235)
Skills are bundle-level domain knowledge, independent of interface type, so they
no longer live under the HTTP-specific pkg/capability. Also add a one-line
conceptual framing (capability = agent-facing projection of an interface) to the
MCP integration docs. No behavior change.
